selinux, xauth, legacy fixes (#6)
* Copy xauthority file to $HOME

If your user has a different UID than 1000, the container user will not
be allowed to read your files.  We use sudo to copy the file to $HOME
instead.  (the sudo ALL rule was not in use previously.)

* support multiple versions/images via different $0 names

also changed the xauth command.  it did not work for me, the
new command looks more obviously correct.

* add support building other old releases